services/user-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.6.0
bcrypt==4.1.2
pyjwt==2.8.0
ddtrace==2.6.0
